# api/docs/swagger.yaml
definitions:
...
  booking.BookingPersonal:
    properties:
      access_status:
        type: string
      count:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: integer
      id:
        type: string
      payment:
        type: integer
      start_date:
        type: string
      subscription_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  booking.ListBookingPersonalResponse:
    properties:
      booking_personal:
        items:
          $ref: '#/definitions/booking.BookingPersonal'
        type: array
    type: object
  booking.ListSubscriptionCoachResponse:
    properties:
      subscription_coach:
        items:
          $ref: '#/definitions/booking.SubscriptionCoach'
        type: array
    type: object
...
  booking.SubscriptionCoach:
    properties:
      coach_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: integer
      description:
        type: string
      duration:
        type: integer
      gym_id:
        type: string
      id:
        type: string
      price:
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /v1/booking-personal:
    get:
      consumes:
      - application/json
      description: Get a list of booking personal records.
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListBookingPersonalResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Booking Personal
      tags:
      - BookingPersonal
    post:
      consumes:
      - application/json
      description: Create a new booking personal with the provided details.
      parameters:
      - description: Booking Personal details
        in: body
        name: booking_personal
        required: true
        schema:
          $ref: '#/definitions/booking.BookingPersonal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/booking.BookingPersonal'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Booking Personal
      tags:
      - BookingPersonal
  /v1/booking-personal/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a booking personal by its ID.
      parameters:
      - description: Booking Personal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Booking Personal
      tags:
      - BookingPersonal
    get:
      consumes:
      - application/json
      description: Get details of a booking personal by its ID.
      parameters:
      - description: Booking Personal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingPersonal'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Booking Personal by ID
      tags:
      - BookingPersonal
    put:
      consumes:
      - application/json
      description: Update an existing booking personal.
      parameters:
      - description: Booking Personal ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated booking personal
        in: body
        name: booking_personal
        required: true
        schema:
          $ref: '#/definitions/booking.BookingPersonal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingPersonal'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Booking Personal
      tags:
      - BookingPersonal
  /v1/subscription-coach:
    get:
      consumes:
      - application/json
      description: Get a list of subscription coach records.
      parameters:
      - description: Filter by gym ID
        in: query
        name: gym_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListSubscriptionCoachResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Subscription Coach
      tags:
      - SubscriptionCoach
    post:
      consumes:
      - application/json
      description: Create a new subscription coach with the provided details.
      parameters:
      - description: Subscription Coach details
        in: body
        name: subscription_coach
        required: true
        schema:
          $ref: '#/definitions/booking.SubscriptionCoach'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/booking.SubscriptionCoach'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Subscription Coach
      tags:
      - SubscriptionCoach
  /v1/subscription-coach/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a subscription coach by its ID.
      parameters:
      - description: Subscription Coach ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Subscription Coach
      tags:
      - SubscriptionCoach
    get:
      consumes:
      - application/json
      description: Get details of a subscription coach by its ID.
      parameters:
      - description: Subscription Coach ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.SubscriptionCoach'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Subscription Coach by ID
      tags:
      - SubscriptionCoach
    put:
      consumes:
      - application/json
      description: Update an existing subscription coach.
      parameters:
      - description: Subscription Coach ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated subscription coach
        in: body
        name: subscription_coach
        required: true
        schema:
          $ref: '#/definitions/booking.SubscriptionCoach'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.SubscriptionCoach'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Subscription Coach
      tags:
      - SubscriptionCoach
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"